Moved top-level logic into a dedicated controller.
