+++ app/config.js
@@ -7,9 +7,11 @@ export default {
   viewportWidth: 22,
   viewportHeight: 11,
   FPS: 60,
-  spritesData: "./data/sprites.json",
-  mapData: "./data/map.json",
-  mapEntitiesData: "./data/entities.json",
+  dataFiles: {
+    sprites: "./data/sprites.json",
+    map: "./data/map.json",
+    entities: "./data/entities.json",
+  },
   spritesDir: "./sprites/",
   waitOnGameStateChange: 3000,
   "debug": true,
